feat: new 404 page with search functionality
* when a user lands on a 404 page, the pathname of the url is used to
create a search term which we query swiftype with to display relevant
results to the user. additionally, the results respect the locale of
the user.
* if no results are found, result section of the page is not shown.
* search input is populated with the search term we create mentioned
in the first bullet. when the user clicks the search input, the search
modal opens.
* we track: the url the user tried to navigate when they got a 404,
how many results are shown to the user, clicks on search results
presented to the user.