fix tests for `useInstrumentedHandler`
`createTessen().{action,track}` are both async functions as of `ed06948`.
jest wasn't waiting for them to finish before checking to see if
`window.Tessen.track` was called, so it thought it wasn't.
this hooks previously only returned the result of the provided callback,
it didn't give any visibility into the result of `window.Tessen.track`.

with this new property `tessenResult`,
we can `await` the Promise in the test suite to make sure it's
finished running before checking to see if `window.Tessen.track` ran