fix: popover closing too early
after changing the font size in `63c1cbef`, the `::after` was too short.
moving your mouse too slowly between the button and the popover
meant it would hide while you were trying to look at it.
this fixes that, and adds a short delay before hiding to give the user
the chance to mouse back over it