fix: codeblocks after LicenseKey not rendering properly
`div` and `p` tags cannot be descendants of `p` tags.
i think normally HTML is fine with this, but i think something with the
MDX parsing was tripping up with it.
changing all the `div` and `p` tags to `span`s isn't pretty,
but it does make codeblocks that come afterwards render properly