Make Link component dynamic based on loggedIn
Added the LoggedInProvider to list of items being exported from the
theme and added that provider to the demo page.

Updated Link component to see if urls are product URLs, if they are, and
the user is not logged it, it will send them to the signup page.  If
they are logged in, it will send them to the product.