Add function to ensure no trail slash in q param
We were having an issue where the 404 page would update the query params
based on the slug that lead the user to get a page that doesn't
exist. Gatsby was adding a trailing slash to this param.  I've added a
function that, if there is a trailing slash in the search term, it will
be removed.  Also, updated the gatsby-config to match the functionality
of the docs site, this enabled me to replicate the bug locally.