style fix: switch from `css` to `styled`
as part of the upgrade to Gatsby 5, the `css` prop wasn't working on
styled components or regular JSX tags, but _only_ in templates.
regular pages are fine.
i spent days trying to figure out wtf was going on and wasn't able to

Liz figured out that switching from the `css` prop to `styled` works,
so instead of toiling away at a heap of config files, we can just do this