fix: tabs not linking on MDX pages
this check worked fine on the index page in the demo site.
but MDX pages have slightly different element types, so this check was
failing, causing the linking to not work at all.
the check isn't really necessary, i was just trying to be extra safe
and it ended up breaking things 🙃